refactor(checkbox): modify to gen better dts
